Added --gzip support
With the --gzip argument, it will add gzip file metadata for you, and you can manually enter timestamp, operating system, file length, and CRC. Sane values are suggested for everything but the CRC (this is because the program doesn't keep an uncompressed state of the data to do the checksum against). There is enough data in the state to collect what the file length would be though, so a sane value is suggested for this. Of course, you can select insane values for any of these.