Bloated Huffman Table
This recipe is an example of the most bloated huffman table possible, it also doesn't encode any useful length-distance pairs; to further discourage proper compression; in other words, this helps you craft to worst possible compression ratio in a block. The recipe lets you decide if this is the last block or not, in the case that you wanted another block after this one.